complete overhaul of the backend
the main addition is the seperation of the writing to the epaper and the
generation of the screenshot. The advantage of this is that the
screenshot can be taken without root, which was sometimes very buggy and
lead to x11 crashing. Also the usage of the xrandr command line utility
was removed and replaced by using the library.